StaticBitmapImageTransform: Do not force copies
The function StaticBitmapImageTransform::ConvertToColorSpace set the
force_copy parameter to be true.

There is a history of load-bearing color space conversions, but this
looks like it was just an unintentional typo. Add a test to make sure
the mistake doesn't happen again.

Bug: 407564084
Change-Id: If3d15842722092895f83615d1602a07ac93dfddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434006
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442881}
NOKEYCHECK=True
GitOrigin-RevId: 10e2357647c7294aec58dca41adcab952b636fdd