webcodecs: Stop reporting VideoFrame's external memory usage to v8
Video frames (especially 8k frames) take a lot of memory and when
a frame is created and reported to v8 it immediately triggers a major GC.
This causes scores of GCs per second and interferes with playback smoothness.

Since we have VideoFrames.close() and don't rely on GC to free VideoFrames
there is no practical benefit in reporting external memory usage.

If VideoFrames.close() hasn't been called we have a console error message when this happens so that developers know when they've made a mistake.

Bug: 407812151
Change-Id: I228a039a5a29642cdb1a8fc5e2c37618813e5f56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432486
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442903}
NOKEYCHECK=True
GitOrigin-RevId: 6c46d465075457ffde0c63e73ba3622008f1e359