blink: Adjust drm test expectations for persistent license
The spec states that "persistent-license" session type support is
optional by key systems. This test should be adjusted so that if the key
system returns unsupported, we skip through the test, but if the session
type is supported, we should verify the logic is working as expected.

spec:
https://www.w3.org/TR/encrypted-media-2/#dom-mediakeysessiontype-persistent-license

Bug: 349649084
Change-Id: I5d87266955f4fa9a9af5cbf66253c87152e9a7e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434033
Reviewed-by: Feras Aldahlawi <feras@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443056}
NOKEYCHECK=True
GitOrigin-RevId: c16c0a9e6f8e3d85e869235a3135e47939b0dc25