Fix the focus controller logic for invokers of all kinds
This CL renames InterestInvokerData to just InvokerData, and adds
the target popover to that data. This allows an invoker to know
which target element it invoked, which in turn lets the focus
controller logic use that, rather than just the existence of a
popovertarget attribute, to correctly manage focus.

This adds GetOpenPopoverTarget() to Element, which retrieves the open
popover target for an element, regardless of the source:
 - `popovertarget`
 - `commandfor`
 - `interesttarget`
 - JS `showPopover({source:foo})`

This CL also takes the opportunity to clean up capitalization for
several things in InvokerData that shouldn't have been camel case,
which unfortunately means lots of files changed.

This CL takes inspiration from two others, and subsumes both:
https://chromium-review.googlesource.com/c/chromium/src/+/6324529
https://chromium-review.googlesource.com/c/chromium/src/+/6357056

Bug: 326681249
Change-Id: I25de9c68e8206cfaf505f257291d3a7b102d34a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432176
Reviewed-by: Keith Cirkel <chromium@keithcirkel.co.uk>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Keith Cirkel <chromium@keithcirkel.co.uk>
Cr-Commit-Position: refs/heads/main@{#1443065}
NOKEYCHECK=True
GitOrigin-RevId: de6eb169a3607c04e91aea81b478e58c74966083