[//cc] Rip out kUnpremultiplyAndDitherLowBitDepthTiles feature
No raster buffer provider does unpremultiply-and-dither now, and we're
not planning to resurrect this functionality. Great insight from
vasilyt@:

"Skia doesn't support rendering to unpremultiplied destinations (and
we'd need really compelling reason to implement this), modern APIs don't
support 4444 (or at least not widely) and I'm planning to remove 4444
path all together.

On Android (if it matters anywhere, it would be android) it already
enabled only on Android O and P. On Desktop it's not enabled
with graphite and it affects desktop devices with 512MB ram which is
probably not really what we're targeting."

Bug: 40042400, 403372453
Change-Id: Ic689515cc26b953f62412fc06ebdea0f36d22dc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433603
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443309}
NOKEYCHECK=True
GitOrigin-RevId: 2a66cee3dc5ceb9a04c5a63da266cf007d544f69