Early abort in Offscreen's GetOrCreateRP if provider is invalid
This change aligns the behavior of OffscreenCanvas with that of
HTMLCanvasElement. If the GPU context is lost in the middle of a task,
the provider may become invalid, but the canvas will only be notified
later, when the NotifyGpuContextLost callback is invoked in a following
task. Thus, for the rest of that task, the provider is unusable and
can't be recreated, even though the canvas context isn't yet lost. In
this state, we can early abort GetOrCreateResourceProvider.

Bug: 346766781
Change-Id: Ie6bcd66ed775642dfba624d969009443cfede5da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433431
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443534}
NOKEYCHECK=True
GitOrigin-RevId: c9f11eabf2e1340444cd3e47b7e61f6efe9104bd