Allow comments, <selectedcontent> as descendants of <select>, <div>
This CL adds a check to return early if a descendant of a <select> is a
comment node. Additionally, we now also check for the button's content
model in `TraverseAncestorsAndCheckDescendant`, which is called when a
parent node is either a <div> or a <span>. This fixes the case where a
<select> with a <button> that contains a <div> and the <div> contains
<selectedcontent> would log a console message about the content model
being broken. Tests are added to validate behavior.

Bug: 408118042
Change-Id: I1e1dd9982c5d56a281ea84eb37044fc67a367500
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433007
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1443751}
NOKEYCHECK=True
GitOrigin-RevId: 709950c45b8aaa36205325beb714a075e76e5798