Fix renderer to correctly display 90m and 270m video files
Recently VideoTransformation was fixed to label the rotation and
mirroring metadata of a video file, but lead to an underlying bug where
the renderer did not correctly handle mirrored videos in the case of 90
rotation and 270 rotation. This change allows the renderer to correctly
handle these cases.

Bug: 408830833
Change-Id: Ic1164d4dbc7ab01ff8c353f2b5bce63723e3a539
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440068
Reviewed-by: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Commit-Queue: Syed AbuTalib <lowkey@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443838}
NOKEYCHECK=True
GitOrigin-RevId: 658c7199d664521c1fa2ae9042ec328c2917e3be