Prefetch: Make prefetch_with_subresources test more stable
SpeculationRuleSetTest.PrerenderIgnorePrefetchRules requires
SpeculationRulesPrefetchWithSubresources is disabled (not
SpeculationRulesPrefetchProxy mentioned in the implementation comment),
but the requirement was not ensured before this change. This worked well
as the feature is disabled by default, but it should be better to
explicitly disable it. This CL does it.

Also, this CL renames Prefetch to PrefetchWithSubresources to more
clarify the target rule type.

Bug: n/a
Change-Id: If66562b0bef4c4ce155ce5bbd382c6bf35aa0331
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6439983
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443910}
NOKEYCHECK=True
GitOrigin-RevId: 8e0ac590b54e8bc1c0fc17851b47b8b5e10d7505