Avoid creating a ContextProviderWrapper if acceleration is disabled
In HTMLCanvasElement::ShouldAccelerate, the code has a comment
explaining that SharedGpuContext::ContextProviderWrapper should be
created at the very end of the function, to avoid creating a GPU
context if none is needed. In CL https://crrev.com/c/4518392, a new
condition was incorrectly added to the function after the call to
ContextProviderWrapper, making it possible for the code to create a
ContextProviderWrapper even if it's never used. This CL moves the
wrapper creation to the end of the function, just before it's used.

Bug: 346766781
Change-Id: I0fcb9f0ec45a8abec46daee330786e7f483fdb05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6437916
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444145}
NOKEYCHECK=True
GitOrigin-RevId: 80a2f55c14968ddb528e95e9c996976136addc6e