Get rid of whitespace_ignored_map_.
We can just as well store this information on the LayoutText object
itself. We don't seem to actually fill this that often, but we do
call erase() a bunch of times on the map, so it's nice to get rid
of that; if nothing else, as a simplification.

Speedometer3 with accessibility enabled (Pinpoint M1, LTO but no PGO,
significant results at 99% CI only):

  TodoMVC-React-Complex-DOM                   [ +0.1%,  +1.1%]
  TodoMVC-jQuery                              [ +0.1%,  +1.1%]
  TodoMVC-JavaScript-ES5                      [ -1.2%,  -0.3%]
  TodoMVC-Svelte-Complex-DOM                  [ -2.1%,  -0.3%]

  Score                                       [ +0.0%,  +0.3%]

Change-Id: Icd1101c1802d6c659a1dd1baf8a34a78ac2cf554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6438836
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444205}
NOKEYCHECK=True
GitOrigin-RevId: e44d4e9e58fff193bcfa609c2a060b3d68408b21