Remove read back when downgrading 2D canvas to CPU
When downgrading a canvas from GPU to CPU, we need to rasterize pending
draw ops and initialize the new canvas resource provider with this
raster. Originally, this was done by drawing the raster into the new
recording.

CL https://crrev.com/c/2702841 fixed a bug where printing would fail if
the recording held a GPU raster. Thus, when downgrading a canvas to
CPU, the GPU raster produced from the pending paint ops would need to
be converted to a CPU raster before it's saved in the new recording.

In CL https://crrev.com/c/4903850 however, the restore logic was
changed to directly restore the pixel in the resource provider via
RestoreBackBuffer, instead of by adding a drawImage paint op. It turns
out that RestoreBackBuffer already reads back the pixels from the GPU
texture [1], so there is no need to also do it in
ReplaceExistingResourceProviderFor2DContext.

[1]: https://crsrc.org/c/third_party/blink/renderer/platform/graphics/canvas_resource_provider.cc;l=2039;drc=944de1047ef3bc5cd8925808bb6109abd9209590

Bug: 346766781
Change-Id: I12737b20319ef4b9e79f5b1289f404111c4e8e05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440053
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444326}
NOKEYCHECK=True
GitOrigin-RevId: 47411a462ce17f7f6c949f8ddeddf403d4a5bcec