[blink] Fix WindowProperties initialization without snapshot
When running a build with `use_v8_context_snapshot = false`, the
`WindowProperties` object doesn't end up properly initialized. This
causes the resulting object to have an undesired `constructor` property
that breaks the prototype chain.

When running with the snapshot enabled, this bug doesn't occur due to a
side-effect of the snapshot hydration. Getting the prototype of the
`Window` interface object is enough to trigger this initialization.

The fix is to ensure that even when the parent has
`is_skipped_in_interface_object_prototype_chain` set to `true`, the
direct parent constructor still gets initialized. That way the
`WindowProperties` type is properly initialized regardless of the
initialization flow.

Bug: 374947853
Change-Id: I71f96f8a842d012978b4286c956b9fb683c5cfe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432174
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444537}
NOKEYCHECK=True
GitOrigin-RevId: c542bf37f38e1e5fdfcaa25baba74d4022d66ff6