[Blink] Move BaseRenderingContext2D away from SkColorType
BaseRenderingContext2D used to not inherit from CanvasRenderingContext,
which meant that it had to go through CanvasRenderingContextHost to
obtain its SkColorType [1]. It now inherits from
CanvasRenderingContext \o/, so it no longer needs to loop through
the host to get this information.

Notes:

* I've left the checks on the presence of `host` present in this CL for
it to be a behavioral no-op. I'll change those as suitable in followup.
For this CL: Both HTMLCanvasElement and OffscreenCanvas clear the
context's pointer to them at the same time as they clear their pointer
to the context, so if the "has host" check succeeds it is guaranteed
that `GetCanvasRenderingContextSkColorType()` will in fact be calling
through to the context's `GetSharedImageFormat()` method (and hence this
CL is a behavioral no-op).

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/canvas_rendering_context_host.cc;l=367-375?q=GetRenderingContextSk&ss=chromium

Bug: 371227617
Change-Id: Ib93aebd5c1faa0c1deaf6c9876d76a8c1f63b2e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6441166
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444560}
NOKEYCHECK=True
GitOrigin-RevId: a79f649a2681da7aafb7bcca03159716c03562b8