[DIP] Fix flaky test
This CL updates the DocumentIsolationPolicy shared-workers WPT test,
that was flaky. The test was doing several calls for workers in parallel
with different configurations, and sometimes one would resolve before
the other, with a different result from the baseline. It was
unnecessary to load all those workers in parallel, so now each subset
only loads one worker. In addition, the worker is created from a remote
executor context, which allows to experiment with different values of
DIP for the worker creator. The previous alternative created an iframe
and reached into its DOM directly to create the worker, but this doesn't
work if the iframe and the main frame of the test have different DIP.
Finally, we removed tests that tried to create workers from FileSystem
URLs because SharedWorkers cannot be loaded from FileSystem URLs.

Bug: 333029146
Change-Id: Iaf74a8efb1268d7b44424beda9531b099b62f179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440923
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Camille Lamy <clamy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444604}
NOKEYCHECK=True
GitOrigin-RevId: 379c4fb91f7329622c9d3135e099df5dcc3c5a57