[animation-trigger] Support auto for animation-trigger-exit-range
animation-trigger is a shorthand that specifies, among other
properties, 4 points that represent the boundaries at which the trigger
will act on its animation(s). The 4 points correspond to the CSS
properties:

 * animation-trigger-range-start
 * animation-trigger-range-end
 * animation-trigger-exit-range-start
 * animation-trigger-exit-range-end

The first 2 properties define the "trigger range" and the last two
define the "exit range." animation-trigger-exit-range is the shorthand
for the last two properties. The default value for the exit range
boundaries is "auto" which can mean one of two things:

1. Primarily,the meaning of auto is "match the trigger range."
2. In the case of animation-trigger-exit-range-end, if a named range,
   e.g. "contain" has been specified for
   animation-trigger-exit-range-start, then auto should be interpreted
   as the end of that named range, similar to normal for
   animation-range.

The slightly expands the meaning of normal so that the case of
"auto normal" has the same meaning as "auto auto", i.e. "match the
trigger range."

Note: This has not yet been resolved on by the CSS working group but it
is being discussed[2] and is how we're thinking about how it all should
work.

[1] https://drafts.csswg.org/css-animations-2/#animation-trigger-exit-range-start
[2] https://github.com/w3c/csswg-drafts/issues/11932

Bug: 390314945
Change-Id: I3a76401e0bf40bebe2960b96d11ba96bb010e705
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426886
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444687}
NOKEYCHECK=True
GitOrigin-RevId: 70e41d785332173843e33681fdebecfc93968e80