[SDA] Fix race condition in update of pending animations
Scroll-driven animations get their deferred start time post layout.
Previously, PreCommit would cancel and restart an animation if missing
a start-time; however, this is not necessary since the animation is
marked as compositor pending once the start time is set. The change
to remove the cancel-restart exposed a race, where we can notify the
animation is ready before it has a start time. This results in DCHECK
failures if enabled, and an idle animation stuck in limbo otherwise.
These animations can simply be dropped from the pending list as they
be picked up again as soon as the start time is resolved.

Bug: 409086867
Change-Id: I70bfd99db6ea6f27aab19c9c30aa0ccf7ff0cf73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440056
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444704}
NOKEYCHECK=True
GitOrigin-RevId: 7d712c2ae79d9ed4a119963c1582fbabf71e397b