Remove raster mode check in ResetLayer
This is a very old check, originally introduced in 2016 by
https://codereview.chromium.org/2123623002. Prior to that CL,
Canvas2DLayerBridge only had an m_layer if the canvas was accelerated
and that layer would be cleared when the bridge was destroyed. That CL
added Canvas2DLayerBridge::disableAcceleration(). This method would
change the raster mode and clear the layer, but it would not delete it.
In this state, the layer didn't need to be cleared again when the bridge
was destroyed because that was already taken care of. Raster mode checks
were added to differentiate valid m_layer used for acceleration canvases
from disabled m_layer left over when disabling acceleration.

16 days later, still in 2016, https://codereview.chromium.org/2171513003
was submitted, deleting Canvas2DLayerBridge::disableAcceleration().
This presumably made the raster mode check obsolete.

Fast forward to 2023 and https://crrev.com/c/4296201 enabled direct
compositing of CPU buffers. Before that CL, we would only create a
cc::Layer if the CanvasResourceProvider was accelerated. Now, we create
a cc::Layer if the CanvasResourceProvider supports direct compositing.
Since CanvasResourceProviderSharedImage in CPU mode supports direct
compositing, we now create cc::Layer in cases where GetRasterMode()
return RasterMode::kCPU. Thus, we cannot clear the layer conditionally
on the raster mode in CanvasResourceHost::ResetLayer().

Bug: 346766781
Change-Id: I04a3b16890bb8c4356f34efe85b06c656b9cf4b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440085
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444718}
NOKEYCHECK=True
GitOrigin-RevId: ff1ca2e75c6a101eed1e52227510d81e5ad079f6