[gnrt] Vendor a real crate when a dependency is no longer trimmed.
Changing which crate features are enabled can change whether a given
crate is present in the resolved dependency tree.  In particular, it may
require deleting a "placeholder" crate and replacing it with a freshly
downloaded, real crate.  This wasn't happening before this CL, because
`gnrt` would check that `foo-vN` exists and contains `Cargo.toml` for
the right version of a crate.  After this CL, `gnrt` also uses some
heuristics to check if `foo-vN` contains a real vs a placeholder crate.

Bug: None
Change-Id: I07517cc45506a45252b785b8e5f616a25e16c421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426604
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1444740}
NOKEYCHECK=True
GitOrigin-RevId: 3a12cd466e40aa9f9916a759202942e7c2a1b554