Support virtual tests for Android and WebView
Developers has expressed there is a need to run WPTs with different
feature set on Android. As we are always using smoke test list for
Android, this change make it possible to run virtual tests on Android.

To run a VTS on swarming bots, we need to update the platform config to
include Android, and add such tests to android.filter.

Bug: None
Change-Id: I1224076e8c1346fb7a63569372a9d38a56856737
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6441778
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1444927}
NOKEYCHECK=True
GitOrigin-RevId: d7ebaa9c1c241e00d5dc4fc2f7af6951f8a43b02