corner-shape: remove curvature adjustment
Curvature adjustment created the illusion that the border had
consistent thickness.

However, the thicker the border, the closer the inner curvature came
to scoop/straight, and stopped resembling the original shape.

Undoing the curvature adjustment - as having a bit of a belly in the
very high/low curvatures is better than not having an inner shape.

Bug: 399449172
Change-Id: I84da4f36b10edbf1fa3f3f8a3b6270e5dc9c440f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6443779
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1445212}
NOKEYCHECK=True
GitOrigin-RevId: 8b86d7fb46ce33bb3932c8dcc7dc7c998ac2d729