SpeculationRules: Add WPT for deduped and sorted speculation tags within rule level
This is a follow-up of https://crrev.com/c/6438143.

Unlike the test added by the previous CL, the new test case checks if
tags are deduped and sorted when they are specified only within rule
level (no tags on ruleset level). In this case, no tags on one of the
rules should result in the null tag.

Bug: 404294138
Change-Id: I1c215845fa40b3b25e6bdf1786bab1cb155b3227
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440785
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445633}
NOKEYCHECK=True
GitOrigin-RevId: 8d7c8e51241063fd0b88ef6d8bde4223ee0ca423