Fix encoded_body_size for the race-network-and-fetch-handler
This is the bug fix CL to address the issue that the `encodedBodySize`
in the resource timing API has `0` even if it actually has body data.

This is because the response head lacks `was_fetched_via_service_worker`
flag when the matched router source is `race-network-and-fetch-handler`
and the network wins the race. Without `was_fetched_via_service_worker`,
the encoded body size is not calculated in `DocumentLoader` [1].

This is a short-term fix, eventually we should change the code to handle
the response from the network as if the response came from the fetch
event for the case of `race-network-and-fetch-handler` for the
consistnecy.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/loader/document_loader.cc;l=1378-1384

Bug: 404577046
Change-Id: I400a2563fc957a97053b98e64b28048b63cdcbf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432968
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445641}
NOKEYCHECK=True
GitOrigin-RevId: ce7760e7ed8287077e074ce55a23fc1d9281c8a0