Use the bitmap acceleration instead of raster mode in ShouldApplyNoise
The CanvasInterventionsHelper was using the raster mode of the canvas
to determine whether noise should be added. The raster mode was
queried using CanvasResourceHost::GetRasterMode(). This function uses a
somewhat convoluted heuristic to determine the raster mode [1]. In
particular, it tries to predict what the raster mode would be if we do
not yet have a resource provider. This can happen if the canvas is
newly created, or if the GPU context was lost. Moreover,
HTMLCanvasElement and OffscreenCanvas have different heuristics in
these cases and will possibly return different values.

It seems that a far simpler and more robust approach would be to
directly check whether the snapshot bitmap is accelerated. The bitmap
is accelerated if and only if the canvas itself is accelerated.

[1]: https://crsrc.org/c/third_party/blink/renderer/platform/graphics/canvas_resource_host.cc;l=100;drc=352fbdae1017811d80b9d0d10a28464e24fae395

Bug: 346766781
Change-Id: Ic209d8498b19c91170ab58bcf3e50842fc449242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440837
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Tom Van Goethem <tov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445674}
NOKEYCHECK=True
GitOrigin-RevId: 0b221257fd7e14da5ceea5e9b97cad3e9deecd9a