[Blink] Rename FrameResource::canvas_resource for clarity
This variable holds the ref used by the placeholder canvas and is set
only when either (a) the placeholder canvas has sent back that ref or
(b) we weren't able to send it to the placeholder canvas in the first
place. Rename it to `placeholder_resource_ref` for clarity.

Bug: 352263194
Change-Id: I73fe90d51b04937e05919d7dd6e3e85446dec7fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6439645
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445706}
NOKEYCHECK=True
GitOrigin-RevId: d8cbc87ccc81bdfbd64ef65c339340168012511b