[Blink] Fix CanvasResourceDispatcher test resource lifetime handling (2)
It should ideally return the placeholder's ref on the CanvasResource
back to the CanvasResourceDispatcher rather than letting it drop on the
floor as it does now, which breaks the invariant that the placeholder
canvas returns its ref to the CanvasResourceDispatcher when it's done
with it.  Upholding this invariant in these tests will be necessary for
followup work I'm doing (which is what uncovered this issue).

For convenience of code organization in the test being changed here I
just ensure that the dispatcher is destroyed before the CanvasResources
in question are dropped.

Bug: 352263194
Change-Id: I796aa523fa189c7078c55d681855f2ef47ceac78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6448309
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445717}
NOKEYCHECK=True
GitOrigin-RevId: 20ccfb3377f5bdca04fcf843193ffe91463e0124