Test provider acceleration in CanvasHibernationHandler::Hibernate
CanvasHibernationHandler::Hibernate was using GetRasterMode() to
validate that the canvas is GPU-accelerated before hibernating it. The
values returned by this getter are hard to interpret:
 - For healthy CPU canvases, return CPU
 - For healthy GPU canvases, return GPU.
 - If the canvas is hibernating, return CPU.
 - For canvas that never had a provider (newly created), return CPU.
 - For canvas that had a provider but lost it (context lost or resize):
   - If the canvas is an OffscreenCanvas, return CPU
   - If the canvas is an HTMLCanvasElement:
     - If preferred_2d_raster_mode_ is CPU, return CPU.
     - If preferred_2d_raster_mode_ is GPU:
       - If the GPU process was revived before the canvas context was
         restored, return GPU.
       - If the GPU process is still down, return CPU.

With this in mind, every time we use GetRasterMode(), we need to wonder
if our logic is compatible with all the above. Everything works in
Hibernate, because we previously checked, directly or indirectly, that
we were not yet hibernating (though only with a DCHECK), that we had a
provider, that it was valid and that the context wasn't lost. Given
these preconditions, calling GetRasterMode() is equivalent to calling
IsAccelerated() on the provider. Calling IsAccelerated() directly
simplifies the logic and reduces cognitive load on the reader.

Bug: 346766781, 409576869
Change-Id: I56cd6659a3a8b0fc7e5fa643a0ed6b34bd752339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6448619
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1445962}
NOKEYCHECK=True
GitOrigin-RevId: 7470f60960f031d8dd8156310a8fd64c9a1977e9