Preserve infinite hit test areas when hit-testing through transforms
In LayoutSVGForeignObject::NodeAtPointFromSVG() an infinite hit test
area is specified to PaintLayer::HitTest(). If the content of the
foreignObject has a transform with a scale which inverse will cause the
"infinite rect" to "grow" and be shifted, it can eventually be shifted
out far enough to no longer intersect the content.

If the hit test area is the "infinite rect", then keep that instead of
using the mapped bounds of it.

Fixed: 408442766
Change-Id: Ibcddbfdbd5abc2d7ff22c94f631e5608a8d6e8e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6440213
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446434}
NOKEYCHECK=True
GitOrigin-RevId: e68d904ae90b0bef190683d376487aac68ed2498