Don't allow interactivity to escape inertness
Let inertness set by interactivity be forced for the subtree using the
same flag as for html inertness caused by modal dialogs and fullscreened
elements.

Also, let ::scroll-marker have a used inertness from its
::scroll-markers per resolution in [1].

The changed expectations in the inert reasons in the AX unit tests are
due to the fact that we now cannot see from the IsInert() flag whether
it was also set due to a modal dialog when the interactivity property
is set to 'inert' at the same time.

[1] https://github.com/w3c/csswg-drafts/issues/11746#issuecomment-2773041391

Bug: 370065759

Change-Id: I34b96698f1157b6fc49cce29daeb890114e58e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6441921
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446485}
NOKEYCHECK=True
GitOrigin-RevId: 283387d34c990502d156a650692c5b9f7d79c272