Avoid repainting when svg image view is not really changed
If the SVGImageElement reports a change in the length attributes
in `SVGImageElement::SvgAttributeChanged`, and the x,y,width,height
attributes are css properties we would like to invalidate paint for
this image only if the viewport has changed since that's only when
we can be sure if the image view has really changed for the user.
However, in the SVGImageElement we do not have the ComputedStyle
available to make the decision on whether a viewport change has been
there or not.
Thus, we choose to do this invalidation whenever needed in the
layout class instead i.e. LayoutSVGImage here whenever a
style-change is made by checking the necessary properties because
there we have the information as to what exact style properties
have been changed.

Bug: 41162195
Change-Id: I0757db39adda1a0e787c8d63844bafba39d55f65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6441026
Commit-Queue: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1446635}
NOKEYCHECK=True
GitOrigin-RevId: 6b18eecf305be9ff4c557dc036d140748efd5fb6