[gap-decorations] Address bug for GapDecorations in column flexbox
This CL addresses issues with the implementation of GapDecorations in
column flexboxes. Previously, the algorithm for computing the
`GapIntersection`s was computing them as (`main_offset`, `cross_offset`)
which breaks down for a column flex box since the main axis offset of an
item becomes a block offset, and a cross axis offset for an item becomes
an inline offset. You could also think about it as, a main axis gap
becomes a column when in a column flexbox.

To solve this, when we are about to add the final intersection to the
vectors of intersections we keep, we "convert" the intersection to be
(inline, block) if needed. So if it is a column flexbox, we convert from
(main_offset, cross_offset) to (cross_offset, main_offset).

This CL also renames some variables and adds context to make the code
and this fix more clear.

Lastly this CL adds a `ToString` method for GapIntersections as a
helpful method for debugging.

Bug: 357648037
Change-Id: I830e343237f3b2972d46d8c73ace7a2e0df9aa6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6405592
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1446678}
NOKEYCHECK=True
GitOrigin-RevId: 2ce8f2091225114832b4cc4649441f3ac3e86fd1