Stop aggressively discarding resources when canvas becomes visible
When a page with a 2d canvas is hidden, the GPU context is configured to
aggressively discard resources. When the page becomes visible, the
context is supposed to resume normal operations. Because this is
specific to GPU accelerated canvas, the logic was behind a GetRasterMode
check. GetRasterMode however always returns kCPU during hibernation,
because while hibernated, the canvas is effectively a usable CPU image
source.

This CL removes the raster mode check. The aggressive resource
discarding is instead only done if we have a GPU context.

Bug: 409576869
Change-Id: I6914b5a38fb3a0f6254e1b7fe2ea7f9110da69f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6448626
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446693}
NOKEYCHECK=True
GitOrigin-RevId: 769c410ea835792d6fd70376a37199623683ba50