Don't use GetRasterMode in HTMLCanvasElement::UpdateMemoryUsage
HTMLCanvasElement::IsAccelerated() calls GetRasterMode, and as
explained in https://crbug.com/409576869, GetRasterMode is convoluted
and hard to interpret. GetRasterMode's return value depends on whether
the canvas is hibernating, whether the canvas is an HTMLCanvasElement or
an OffscreenCanvas, whether it ever had a provider or whether it lost
it, etc. Instead of using GetRasterMode, callers are better off being
more explicit with what they care about. Here, what we really want is
check if the provider is accelerated.

Bug: 409576869
Change-Id: I144042d2083fc106729f7162e429b8d701994325
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6454504
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446695}
NOKEYCHECK=True
GitOrigin-RevId: c91d94e1acb596adaf89f38ff3a7b4c20c7b8c9d