Support corner-shape with backdrop-filter
The backdrop filter bounds clip the backdrop without
clipping the element itself.
With corner-shape (and later with border-shape), those
can be an arbitrary path, not necessarily an RRect.

Pass around SkPath instead of gfx::RRect all the way
from PaintLayer down to the GPU process for the backdrop
filter bounds, and use that to clip the filter instead
of the RRect, keeping the existing optimizations when
the path can be simplified back to an RRect.

Bug: 402437852
Change-Id: I51bd39a47499c2fbf3f4628627f771677658f9ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6437734
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1446714}
NOKEYCHECK=True
GitOrigin-RevId: 77aa3dc409b6fd0605ee4ceb4eb103436c1d7c69